+++ config/jetstream.php
@@ -49,4 +49,17 @@ return [
         Features::accountDeletion(),
     ],
 
+    /*
+    |--------------------------------------------------------------------------
+    | Profile Photo Disk
+    |--------------------------------------------------------------------------
+    |
+    | This configuration value determines the default disk that will be used
+    | when storing profile photos for your application's users. Typically
+    | this will be the "public" disk but you may adjust this if needed.
+    |
+    */
+
+    'profile_photo_disk' => 'public',
+
 ];