+++ resources/views/components/form-section.blade.php
@@ -8,19 +8,17 @@
 
     <div class="mt-5 md:mt-0 md:col-span-2">
         <form wire:submit.prevent="{{ $submit }}">
-            <div class="shadow overflow-hidden sm:rounded-md">
-                <div class="px-4 py-5 bg-white sm:p-6">
-                    <div class="grid grid-cols-6 gap-6">
-                        {{ $form }}
-                    </div>
+            <div class="px-4 py-5 bg-white sm:p-6 shadow {{ isset($actions) ? 'sm:rounded-tl-md sm:rounded-tr-md' : 'sm:rounded-md' }}">
+                <div class="grid grid-cols-6 gap-6">
+                    {{ $form }}
                 </div>
-
-                @if (isset($actions))
-                    <div class="flex items-center justify-end px-4 py-3 bg-gray-50 text-right sm:px-6">
-                        {{ $actions }}
-                    </div>
-                @endif
             </div>
+
+            @if (isset($actions))
+                <div class="flex items-center justify-end px-4 py-3 bg-gray-50 text-right sm:px-6 shadow sm:rounded-bl-md sm:rounded-br-md">
+                    {{ $actions }}
+                </div>
+            @endif
         </form>
     </div>
 </div>