+++ src/Http/Controllers/Livewire/TeamController.php
@@ -38,6 +38,8 @@ class TeamController extends Controller
      */
     public function create(Request $request)
     {
+        Gate::authorize('create', Jetstream::newTeamModel());
+
         return view('teams.create', [
             'user' => $request->user(),
         ]);