+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/jetstream/compare/v2.1.2...master)
+## [Unreleased](https://github.com/laravel/jetstream/compare/v2.1.4...master)
+
+
+## [v2.1.4 (2021-02-09)](https://github.com/laravel/jetstream/compare/v2.1.3...v2.1.4)
+
+### Changed
+- Allow custom classes on the dropdown class ([#661](https://github.com/laravel/jetstream/pull/661))
+
+### Fixed
+- Fix permissions for team create routes ([#669](https://github.com/laravel/jetstream/pull/669))
+- Fix overflow hidden issue with overlaping elements inside form section ([#668](https://github.com/laravel/jetstream/pull/668))
+
+
+## [v2.1.3 (2021-02-02)](https://github.com/laravel/jetstream/compare/v2.1.2...v2.1.3)
+
+### Changed
+- Add `profile_photo_disk` config option ([#658](https://github.com/laravel/jetstream/pull/658), [a6245c8](https://github.com/laravel/jetstream/commit/a6245c84adc2ea209a34051ff543ccc302a275e8))
 
 
 ## [v2.1.2 (2021-01-26)](https://github.com/laravel/jetstream/compare/v2.1.1...v2.1.2)